+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Used fetch API for downloading fonts and images
+- Update Roboto font (version 3.013)
 
 ## 0.3.0-beta.18 - 2025-05-09