+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.2.21 - 2025-12-22
 
 - Added `wordBreak` property for `text` node, supported values: `'normal'` (default), `'break-all'`
 - Update Roboto font (version 3.014)